Change charger current from a uint16 to a uint8 (maximum is 32 Amps for most regions or 80 Amps for some chargers in split voltage countries)

+++ io.h
@@ -35,7 +35,7 @@ class TeslaControllerIO {
         virtual void writeChargerCurrent(uint16_t, uint8_t, uint8_t) = 0;
         virtual void writeTotalConnectedChargers(uint8_t) = 0;
         virtual void writeChargerFirmware(uint16_t, EXT_FIRMWARE_PAYLOAD_T*) = 0;
-        virtual void writeChargerActualCurrent(uint16_t, uint16_t) = 0;
+        virtual void writeChargerActualCurrent(uint16_t, uint8_t) = 0;
         virtual void writeState() = 0;
         virtual void stopCharging() = 0;
 };
